requirements.txt
Flask
python-dotenv
pytesseract
Pillow
requests
PyPDF2
PyMuPDF
opencv-python
numpy 